games stats swapped to displaying game level instead of total number of completed quests
